Support client specification for request bodies.
This commit adds support for consumers of an OkHttp Request.Builder to
override the default request body handling.  This is useful for talking
to servers that might consider a header frame with an unfinished flag an
invalid request, e.g. a server that doesn't think DELETE should have a
request body.

Jira-Id: ACCTSEC-4720
Change-Id: Id72201083e524d55b4e8a2810565011d1401d497