Add a method to explicitly mark connections as stale.
This change adds a method that allows clients to inform the connection
pool that the existing connections are stale. It does not explicitly
close the connections, only marking them as unusable for future HTTP
streams.

The specific use-case here is when the client knows that the network
link has changed and doesn't want the network to wait for timeouts
before it discovers the connections are closed.

Jira-Id: ANDROID-23267
Change-Id: Iae0d8f5b3e51b2a517275f16f1ad31b48a32f228